requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.2
python-dotenv>=1.0.0

# Web Panel Dependencies
Flask>=3.0.0
Flask-Session>=0.5.0
requests>=2.31.0
requests-oauthlib>=1.3.1

# Template Engine
Jinja2>=3.1.2

# Utilities
aiohttp>=3.9.0
pytz>=2023.3
